add hibernate, many to many does`t work

+++ core/src/main/java/com/epam/esm/dto/CustomPage.java
@@ -11,11 +11,11 @@ import java.util.Collection;
 public class CustomPage<T> extends RepresentationModel<CustomPage<?>> {
     private Collection<T> content;
     private Integer size;
-    private Integer totalElements;
-    private Integer totalPages;
+    private Long totalElements;
+    private Long totalPages;
     private Integer number;
 
-    public CustomPage(Collection<T> content, CustomPageable pageRequest, Integer total) {
+    public CustomPage(Collection<T> content, CustomPageable pageRequest, Long total) {
         this.content = content;
         size = pageRequest.getSize();
         totalElements = total;
